+++ scripts/changelog.ts
@@ -185,7 +185,9 @@ async function exec(): Promise<void> {
         '**core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty',
         '**core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition',
         '**plugin-text**. Added visual latex editor',
-        '**plugin-text**. Added headings plugin'
+        '**plugin-text**. Added headings plugin',
+        '**plugin-text**. Added list plugin',
+        '**plugin-text**. Added colors plugin'
       ],
       changed: [
         '**plugin-text**. Hide placeholder in render mode',
@@ -194,11 +196,7 @@ async function exec(): Promise<void> {
         '**ui** / **editor-ui** / **renderer-ui**. Revised theming workflow',
         '**ui**. reworked the ui of several plugins (image, video, anchor, geogebra, h5p, highlight, hint, solution and spoiler)',
         '**plugin-rows**. Improved styling of add menu and controls',
-        '**plugin-rows**. Added more theming options',
-        '**core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty',
-        '**core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition',
-        '**plugin-text**. Added visual latex editor',
-        '**plugin-rows**. Improved styling of add menu and controls'
+        '**plugin-rows**. Added more theming options'
       ],
       fixed: ['**plugin-image**. Hide config-overlay in render mode'],
       internal: [